Change hashtags to headers
At the moment, this file doesn't include a space after a hashtag. According to the markdown specification, at least 1 space is required. Because this space is missing, it means that markdown renderers such as github don't render the file headers, giving this project a bad presentation to the outside world